backend/requirements.txt
django-cors-headers==3.13.0
django-environ==0.9.0
django-jazzmin==2.5.0
Django==4.0.5
djangorestframework==3.13.1
mysqlclient==2.1.0
pandas==1.4.3
psycopg2==2.9.3
